# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Charging Stations API Overview",
        "description": "Charging Stations API documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/stations/{id}": {
            "get": {
                "tags": [
                    "Station"
                ],
                "summary": "Get station details",
                "operationId": "368911e17e985f1cf255154b55ec5c20",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Upton Group"
                                        },
                                        "latitude": {
                                            "type": "number",
                                            "example": "22.123123"
                                        },
                                        "longitude": {
                                            "type": "number",
                                            "example": "132.341233"
                                        },
                                        "address": {
                                            "type": "number",
                                            "example": "8516 Dayne Vista Suite 240 West Oran, ND 22379"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Station"
                ],
                "summary": "Delete station",
                "operationId": "ae4854e3e831ada274aad1d97f5347ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Station deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/stations/search": {
            "get": {
                "tags": [
                    "Station"
                ],
                "summary": "Get stations within radius",
                "operationId": "fdc4b498a4f92e87d684664eba9fbf74",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Latitude",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": "27.90461"
                        }
                    },
                    {
                        "name": "lon",
                        "in": "query",
                        "description": "Longitude",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": "59.83534"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Search Radius (in kilometers)",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": "50.0"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "stations": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "location": {
                                                        "properties": {
                                                            "latitude": {
                                                                "type": "string",
                                                                "example": "27.90461"
                                                            },
                                                            "longitude": {
                                                                "type": "string",
                                                                "example": "59.83534"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "companies": {
                                                        "properties": {
                                                            "5": {
                                                                "properties": {
                                                                    "company_id": {
                                                                        "type": "number",
                                                                        "example": 5
                                                                    },
                                                                    "stations_in_company": {
                                                                        "type": "array",
                                                                        "items": {
                                                                            "properties": {
                                                                                "id": {
                                                                                    "type": "number",
                                                                                    "example": 79
                                                                                },
                                                                                "name": {
                                                                                    "type": "string",
                                                                                    "example": "Kohler-Buckridge"
                                                                                },
                                                                                "company_id": {
                                                                                    "type": "number",
                                                                                    "example": 5
                                                                                },
                                                                                "address": {
                                                                                    "type": "string",
                                                                                    "example": "38369 Imani Circle\\nCarlosstad, UT 85281-5939"
                                                                                },
                                                                                "distance": {
                                                                                    "type": "number",
                                                                                    "example": 1378.317035354452
                                                                                }
                                                                            },
                                                                            "type": "object"
                                                                        }
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}